requirements.txt
hazm==0.10.0
wordcloud==1.9.4
arabic-reshaper==3.0.0
python-bidi==0.6.3
loguru==0.7.3